refactor: extract config screens
